Refactor: Move calendar event code to calendar_events.go
Per PR review, keep calendar-event tooling in its own file so it can grow
as more calendar-related tools are added.
