+++ internal/twprojects/calendars_test.go
@@ -30,14 +30,3 @@ func TestCalendarList(t *testing.T) {
 		"page_size": float64(10),
 	})
 }
-
-func TestCalendarEventList(t *testing.T) {
-	mcpServer := mcpServerMock(t, http.StatusOK, []byte(`{}`))
-	testutil.ExecuteToolRequest(t, mcpServer, twprojects.MethodCalendarEventList.String(), map[string]any{
-		"calendar_id":        float64(123),
-		"started_after_date": "2023-01-01",
-		"ended_before_date":  "2023-12-31",
-		"limit":              float64(25),
-		"cursor":             "abc123",
-	})
-}
